requirements.txt
aiogram~=3.3.0
aiohttp==3.9.3
magic_filter==1.0.12
pydantic_settings==2.2.1
python-dotenv==1.0.1
asyncio~=3.4.3
requests~=2.31.0
DateTime~=5.4
pytz~=2021.3